requirements.txt
mlx-whisper>=0.4.0
pyyaml>=6.0
